Layout_TestCase/http/tests/xmlhttprequest/access-control-allow-lists-starting-with-comma.html

<!DOCTYPE html>
<html>
<head>
<script src="../resources/js-test-pre.js"></script>
</head>
<body>
<script type="text/javascript">
description("Allow lists starting with a comma should be parsed correctly.");
var xhr = new XMLHttpRequest();
var url = "http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests/xmlhttprequest/resources/access-control-allow-lists.php?headers=,y-lol,x-print,%20,,,y-print&origin=http://http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests";
xhr.open("GET", url, false);
xhr.setRequestHeader('x-print', 'unicorn')
xhr.setRequestHeader('y-print', 'narwhal')
debug("Sending GET request with custom headers.");
shouldBeUndefined("xhr.send(null)");
var response = JSON.parse(xhr.response)
shouldBeEqualToString("response['x-print']", "unicorn");
shouldBeEqualToString("response['y-print']", "narwhal");

url = "http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests/xmlhttprequest/resources/access-control-allow-lists.php?methods=,,PUT,GET&origin=http://http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests";
xhr.open("PUT", url, false);
debug("Sending PUT request.");
shouldBeUndefined("xhr.send(null)");
</script>
<script src="../resources/js-test-post.js"></script>
</body>
</html>
